Initial refactoring
Create new baseform and include it in the two upload forms
